+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/services/http/model/GroovyConsoleOutput.kt
@@ -1,3 +1,3 @@
 package com.github.bobi.aemgroovyconsoleplugin.services.http.model
 
-data class GroovyConsoleOutput(val output: String, val runningTime: String, val exceptionStackTrace: String)
+data class GroovyConsoleOutput(val output: String, val runningTime: String, val exceptionStackTrace: String, val table: GroovyConsoleTable? = null)