+++ src/main/kotlin/com/github/bobi/aemgroovyconsoleplugin/services/http/model/GroovyConsoleTable.kt
@@ -1,0 +1,7 @@
+package com.github.bobi.aemgroovyconsoleplugin.services.http.model
+
+/**
+ * User: Andrey Bardashevsky
+ * Date/Time: 06.08.2022 18:51
+ */
+data class GroovyConsoleTable(val columns: List<String>, val rows: List<List<String>>)